% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approx_sb}
\alias{approx_sb}
\title{lower bound of a vector of probabilities that sum to one}
\usage{
approx_sb(x)
}
\arguments{
\item{x}{a vector of probabilities that sum to one}
}
\value{
approximation to a vector of probabilities
}
\description{
the lower bound is based on the best set of local variational parameters
which are logit of the stick-breaking form of the supplied vector
}
\examples{

# based on Tsiatis 2016 NeuroIPS
approx <- function(x){
  res = rep(NA,length=length(x))
  for (i in 1:length(res)){
    curr_v <- x[i] - x[-i]
    res[i]  = prod(expit(curr_v))
  }
  res
}

tau = rep(0.25,4)
barplot(rbind(tau,approx_sb(tau),approx(tau)),beside=TRUE,
        legend.text=c("truth","sb + quad (lotR)","1 vs other + quad"),
        main="truth 1")

tau = c(0.5,0.3,0.15,0.05)
barplot(rbind(tau,approx_sb(tau),approx(tau)),beside=TRUE,
        legend.text=c("truth","sb + quad (lotR)","1 vs other + quad"),
        main="truth 2")

}
\references{
\itemize{
\item Jaakkola, Tommi S., and Michael I. Jordan. "Bayesian parameter estimation via variational methods." Statistics and Computing 10.1 (2000): 25-37.
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.399.9368&rep=rep1&type=pdf}
\item Titsias M(2016). One-vs-each approximation to softmax for scalable estimation of probabilities. Advances in Neural Information Processing Systems.
\url{https://papers.nips.cc/paper/6468-one-vs-each-approximation-to-softmax-for-scalable-estimation-of-probabilities.pdf}
}
}
\seealso{
Other VI functions: 
\code{\link{g_fun.vec}()},
\code{\link{g_fun0}()},
\code{\link{g_fun}()},
\code{\link{lower_bd}()}
}
\concept{VI functions}
