% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcm_tree.R
\name{continue_lcm_tree}
\alias{continue_lcm_tree}
\title{continue from a prevoius model fits}
\usage{
continue_lcm_tree(
  old_mod,
  update_hyper_freq = NULL,
  print_freq = NULL,
  tol = 1e-08,
  tol_hyper = 1e-04,
  max_iter = 5000,
  nrestarts = 1,
  keep_restarts = TRUE,
  parallel = TRUE,
  log_restarts = FALSE,
  log_dir = ".",
  random_init = FALSE,
  allow_continue = FALSE
)
}
\description{
can definitely ignore this function, and use lcm_tree directly with various settings.
So this function is for lazy persons.
}
