% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expit}
\alias{expit}
\title{expit}
\usage{
expit(x)
}
\arguments{
\item{x}{}
}
\description{
expit
}
\seealso{
Other utility function: 
\code{\link{log1p.exp.vec}()},
\code{\link{log1p_exp}()},
\code{\link{logexpit}()},
\code{\link{logit}()}
}
\concept{utility function}
