% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_lcm_tree}
\alias{fit_lcm_tree}
\title{fit the tree-shrinkage latent class models (this function largely follows \code{moretrees})}
\usage{
fit_lcm_tree(
  dsgn,
  vi_params_init,
  hyperparams_init,
  random_init,
  random_init_vals,
  tol,
  tol_hyper,
  max_iter,
  print_freq,
  update_hyper_freq,
  hyper_fixed
)
}
\arguments{
\item{vi_params_init, hyperparams_init, random_init, random_init_vals, tol, tol_hyper, max_iter, print_freq, update_hyper_freq, hyper_fixed}{initial values and updating protocols. Explained more in the wrapper function.}

\item{dat_tree}{a list of data and other information organized according to the tree}
}
\value{
a list with model updates

\describe{
\item{\code{vi_params}}{named list of final variational parameter estimates}
\item{\code{hyperparams}}{named list of final hyperparameter estimates}
\item{\code{hyper_fixed}}{named list of fixed hyperparameters}
\item{\code{ELBO_track}}{numeric vector containing the values of the objective function
(ELBO) at the end of every iteration}
}
}
\description{
Returns the variational parameters (mostly moments of relevant unknown quantities)
}
\concept{internal VI functions}
