% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g_fun.vec}
\alias{g_fun.vec}
\title{g, vectorized}
\usage{
g_fun.vec(eta)
}
\arguments{
\item{eta}{local variational parameter}
}
\description{
g, vectorized
}
\seealso{
Other VI functions: 
\code{\link{g_fun0}()},
\code{\link{g_fun}()},
\code{\link{lower_bd}()}
}
\concept{VI functions}
