% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g_fun0}
\alias{g_fun0}
\title{g}
\usage{
g_fun0(eta)
}
\arguments{
\item{eta}{local variational parameter}
}
\description{
g
}
\seealso{
Other VI functions: 
\code{\link{g_fun.vec}()},
\code{\link{g_fun}()},
\code{\link{lower_bd}()}
}
\concept{VI functions}
