% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_est_cpp}
\alias{get_est_cpp}
\title{Summarize the posterior mean, sd and confidence interval}
\usage{
get_est_cpp(
  node_select,
  mu_gamma,
  sigma_gamma,
  mu_alpha,
  Sigma_alpha,
  anc,
  cardanc,
  z
)
}
\arguments{
\item{node_select}{a vector of zeros and ones, indicating which nodes
are selected based on variational probability (prob > 0.5). Length = p}

\item{mu_gamma}{variational Gaussian means (for \code{s_u=1} component) for J*K
logit(class-specific response probabilities); (J,K,p) array; In R, we used a list of p (J,K) matrices}

\item{sigma_gamma}{variational Gaussian variances (for \code{s_u=1} component)
for J*K logit(class-specific response probabilities); (J,K,p) array}

\item{mu_alpha}{variational Gaussian mean vectors (for \code{s_u=1} component) -
this is a p by K-1 matrix; in R, we used a list of p vectors (each of length K-1)}

\item{Sigma_alpha}{variational Gaussian variances (for \code{s_u=1} component)
\itemize{
\item this is an array of dimension (K-1, K-1, p); in R, we used a list of p matrices,
each of dimension K-1 by K-1.
}}

\item{anc}{a list of pL vectors, each vector has the node ids of the ancestors;
lengths may differ. The ancestors include the node concerned.}

\item{cardanc}{a numeric vector of length pL; integers. The number
of ancestors for each leaf node}

\item{z}{= \code{ ci_level+(1-ci_level)/2}}
}
\value{
a list
\describe{

Named("beta_est")=beta_est,
Named("beta_sd_est")=beta_sd_est,
Named("beta_cil")=beta_cil,
Named("beta_ciu")=beta_ciu,

Named("eta_est")=eta_est,
Named("eta_var_est")=eta_var_est,
Named("eta_cil")=eta_cil,
Named("eta_ciu")=eta_ciu

}
}
\description{
Summarize the posterior mean, sd and confidence interval
}
