% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_line1_2_subid}
\alias{get_line1_2_subid}
\title{calculate line 1 and 2 and 13 of ELBO to assess convergence
and choose among converged estimates from many restarts}
\usage{
get_line1_2_subid(
  psi,
  g_psi,
  phi,
  g_phi,
  rmat,
  E_beta,
  E_beta_sq,
  E_eta,
  E_eta_sq,
  X,
  v_lookup
)
}
\arguments{
\item{psi, }{g_psi see \code{\link{update_gamma_subid}}}

\item{phi}{a (pL,K,K) array of local variational parameters for approximating
the \code{expit(eta_vk-eta_vm)}}

\item{g_phi}{g transformed phi}

\item{rmat}{a matrix of variational probabilities of all observations
belong to K classes; N by K; each row sums to 1}

\item{E_beta, }{E_beta_sq, E_beta_diff, E_beta_diff_sq moments during
VI updates from \code{\link{get_moments_cpp}}}

\item{X}{transformed data: 2Y-1}

\item{outcomes_units}{a list of length pL, each is a vector indicating the
observations in each leaf (aka outcome)}

\item{cardleaf}{a vector of legnth pL; counts the number of observations in each leaf}

\item{idnotk}{utility matrix}
}
\value{
ELBO value; negative
}
\description{
calculate line 1 and 2 and 13 of ELBO to assess convergence
and choose among converged estimates from many restarts
}
