% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_moments_cpp}
\alias{get_moments_cpp}
\title{Get all moments that need update when iterating over a total of p internal and leaf nodes}
\usage{
get_moments_cpp(
  prob,
  mu_gamma,
  sigma_gamma,
  mu_alpha,
  Sigma_alpha,
  anc,
  cardanc
)
}
\arguments{
\item{prob}{variational probabilities for \code{s_u}; length p}

\item{mu_gamma}{variational Gaussian means (for \code{s_u=1} component) for J*K
logit(class-specific response probabilities); (J,K,p) array; In R, we used a list of p (J,K) matrices}

\item{sigma_gamma}{variational Gaussian variances (for \code{s_u=1} component)
for J*K logit(class-specific response probabilities); (J,K,p) array}

\item{mu_alpha}{variational Gaussian mean vectors (for \code{s_u=1} component) -
this is a p by K-1 matrix; in R, we used a list of p vectors (each of length K-1)}

\item{Sigma_alpha}{variational Gaussian variances (for \code{s_u=1} component)
- this is an array of dimension (K-1, K-1, p); in R, we used a list of p matrices,
each of dimension K-1 by K-1.}

\item{anc}{a list of pL vectors, each vector has the node ids of the ancestors;
lengths may differ. The ancestors include the node concerned.}

\item{cardanc}{a numeric vector of length pL; integers. The number
of ancestors for each leaf node}

\item{D_k_array}{a utility array of dimension (K-1,K,K); for each slice,
multiplying it with a vector of length K produces \code{(alpha_k-alpha_m), for m not k}}

\item{idnotk}{K-1 by K, utility matrix, column k indicates the indices that are not k}

\item{submat}{K-1 by K, utility matrix with a K-1-identity, right padded with
a column of zeros.}
}
\value{
a List

\describe{
    Named("E_beta")=E_beta,
    Named("E_zeta")=E_zeta,
    Named("E_beta_sq")=E_beta_sq,
    Named("E_xi")=E_xi,
    Named("E_eta")=E_eta,
    Named("E_eta_sq")=E_eta_sq
}
}
\description{
This is slower than the skinny version (\code{\link{get_moments_cpp}})
}
