% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_moments_cpp_skinny}
\alias{get_moments_cpp_skinny}
\title{get moments that need update when iterating over u (except for \code{rmat})}
\usage{
get_moments_cpp_skinny(prob, mu_gamma, mu_alpha, anc, cardanc)
}
\arguments{
\item{prob}{variational probabilities for \code{s_u}; length p}

\item{mu_gamma}{variational Gaussian means (for \code{s_u=1} component) for J*K
logit(class-specific response probabilities); (J,K,p) array; In R, we used a list of p (J,K) matrices}

\item{mu_alpha}{variational Gaussian mean vectors (for \code{s_u=1} component) -
this is a p by K-1 matrix; in R, we used a list of p vectors (each of length K-1)}

\item{anc}{a list of pL vectors, each vector has the node ids of the ancestors;
lengths may differ. The ancestors include the node concerned.}

\item{cardanc}{a numeric vector of length pL; integers. The number
of ancestors for each leaf node}
}
\value{
a List

\describe{
Named("E_beta")=E_beta,
Named("E_zeta")=E_zeta,
# Named("E_beta_sq")=E_beta_sq,
Named("E_xi")=E_xi,
Named("E_eta")=E_eta,
Named("E_xi_diff")=E_xi_diff,
Named("E_eta_diff")=E_eta_diff,
# Named("E_eta_diff_sq")=E_eta_diff_sq
}
}
\description{
(skinny version of \code{\link{get_moments_cpp}})
}
