% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getC}
\alias{getC}
\title{Initialize sigma alpha for shared tau}
\usage{
getC(u, g_phi, rmat, tau_1_t, h_pau, subject_ids, v_lookup)
}
\arguments{
\item{u, g_phi, rmat, tau_1_t, h_pau, subject_ids, v_lookup}{NB: lazy now, see other functions.}
}
\description{
Initialize sigma alpha for shared tau
}
