% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getC_separate_tau}
\alias{getC_separate_tau}
\title{Initialize sigma alpha for distinct tau}
\usage{
getC_separate_tau(u, g_phi, rmat, tau_1_t, h_pau, subject_ids, v_lookup)
}
\arguments{
\item{u, g_phi, rmat, tau_1_t, h_pau, subject_ids, v_lookup}{see \code{\link[=getC]{getC()}}}
}
\description{
Initialize sigma alpha for distinct tau
}
