% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_tree_lcm}
\alias{initialize_tree_lcm}
\title{Initialize the variational inferential algorithm for latent
class model with tree structured shrinkage}
\usage{
initialize_tree_lcm(
  Y,
  A,
  Z_obs,
  leaf_ids_units,
  leaf_ids_nodes,
  ancestors,
  h_pau,
  levels,
  vi_params,
  hyperparams,
  hyper_fixed,
  random_init,
  random_init_vals,
  subject_id_list,
  v_units,
  shared_tau
)
}
\arguments{
\item{Y}{matrix of binary observations for LCM - rows ordered by leaves in the tree}

\item{A}{a p by p binary matrix: each row is the ancestors including the node itself;
ordered by leaves in the tree}

\item{Z_obs}{a two-column matrix of integers, first column is subject ids,
second column is a mix of NA and integers, NA means unknown class indicators,
an integer indicates the known class.}

\item{leaf_ids_units}{The subject ids in each leaf node (a list)}

\item{leaf_ids_nodes}{the leaf descendants for each internal or leaf nodes (a list)}

\item{ancestors}{a numeric vector of ancestor nodes for each leaf node (a list of length equal to
the number of leaves)}

\item{h_pau}{a numeric vector (length = # nodes); the edge length between a node
and its parent node. The root node has no parent, because we suggest a separate
prior variance for the root node's gamma and alpha, we set the "edge-length" toward root
node as 1.}

\item{levels}{a numeric vector of integers from 1 to L, indicating for each node
(leaf or internal node) which set of hyperparameters to use. For example,
if we want the root node to have a separate \code{tau_1} and \code{tau_2}, can specify it to
its own level. Another example would be to have distinct sets of hyperparameters
for leaf and non-leaf nodes. The levels are pre-specified, not estimated.}

\item{vi_params}{the list of variational parameters. \code{mu_gamma},
\code{mu_alpha}, \code{prob} (for s_u), \code{a_t}, \code{b_t},\code{sigma_gamma}, \code{Sigma_alpha}}

\item{hyperparams}{the list of hyperparameters, \code{tau_1} and \code{tau_2} -
these are initial specifications of the hyperparameters - they are updated by
 \code{tau_1_t}, \code{tau_2_t}; \code{psi}, \code{g_psi}, \code{phi}, \code{g_phi} (these
are not hyperparameters, but are updated with the same schedule as \code{tau_1_t} and \code{tau_2_t}).}

\item{hyper_fixed}{a list of fixed hyperparameters, such as those
in the Beta priors for \code{rho}, e.g., \code{list(a=c(1,1,99),b=c(1,1,1))},
\code{K}, number of classes.
for three levels, where the first and second levels are uniform Beta, the third level
has \code{Beta(99,1)} prior - this nearly ensures setting \code{s_u=1}.}

\item{random_init}{logical; \code{TRUE} for adding additional variability to the
initial values. This is needed if the algorithm needs multiple random starts
to choose the best converged values.}

\item{random_init_vals}{NB: fill out specific elements}

\item{subject_id_list, v_units, shared_tau}{see \code{\link{update_vi_params}}}
}
\value{
a list \code{vi_params,hyperparams} containing the initial values.
}
\description{
Initialize the variational inferential algorithm for latent
class model with tree structured shrinkage
}
\concept{internal function}
