% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{lca_ll}
\alias{lca_ll}
\title{log likelihood for lca}
\usage{
lca_ll(dat, theta_mat, pi_vec)
}
\arguments{
\item{dat}{a binary matrix with rows for observations and columns
for features (of the same dimension as in the training data)}

\item{theta_mat}{a \code{J} by \code{K} matrix of positive response probabilities}

\item{pi_vec}{a \code{K}-vector that sum to one; latent class proportions.}

\item{...}{Other parameters}
}
\value{
log likelihood
}
\description{
log likelihood for lca
}
