% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log1p.exp.vec}
\alias{log1p.exp.vec}
\title{log expit vectorized}
\usage{
\method{log1p}{exp.vec}(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\description{
log expit vectorized
}
\seealso{
Other utility function: 
\code{\link{expit}()},
\code{\link{log1p_exp}()},
\code{\link{logexpit}()},
\code{\link{logit}()}
}
\concept{utility function}
