% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log1p_exp}
\alias{log1p_exp}
\title{log(1+exp(x)) (copied from moretrees:)}
\usage{
log1p_exp(x)
}
\arguments{
\item{x}{}
}
\description{
log(1+exp(x)) (copied from moretrees:)
}
\seealso{
Other utility function: 
\code{\link{expit}()},
\code{\link{log1p.exp.vec}()},
\code{\link{logexpit}()},
\code{\link{logit}()}
}
\concept{utility function}
