% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{sigmoid (copied from moretrees:)}
\usage{
logit(p)
}
\arguments{
\item{p}{}
}
\description{
sigmoid (copied from moretrees:)
}
\seealso{
Other utility function: 
\code{\link{expit}()},
\code{\link{log1p.exp.vec}()},
\code{\link{log1p_exp}()},
\code{\link{logexpit}()}
}
\concept{utility function}
