% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotR.R
\name{lotr}
\alias{lotr}
\title{Lord of the Rings soundboard}
\usage{
lotr(x = "over")
}
\arguments{
\item{x}{Clip. Defaults to "over".}
}
\description{
Plays a quote from LotR upon execution
}
\examples{
lotr("mellon")
}
\keyword{Lotr}
