% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotR_example_data.R
\docType{data}
\name{lotR_example_data_tree_K2}
\alias{lotR_example_data_tree_K2}
\title{Simulated multivariate binary data. (with tree structure)}
\format{
A named list with the following elements:
see the returned values of \code{\link[=simulate_lcm_tree]{simulate_lcm_tree()}}
}
\usage{
lotR_example_data_tree_K2
}
\description{
Simulated data for testing the \code{\link[=lcm_tree]{lcm_tree()}} function.
Used in the example. (K=2)
}
\seealso{
\code{\link[=simulate_lcm_tree]{simulate_lcm_tree()}}
}
\keyword{datasets}
