% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr_palette.R
\name{lotr_palette}
\alias{lotr_palette}
\title{Access LOTR Color Palettes}
\usage{
lotr_palette(name, n = NULL, type = NULL, reverse = FALSE)
}
\arguments{
\item{name}{Name of the palette. Available palettes: "shire", "mordor",
"rivendell", "minas_morgul", "fellowship", "aragorn", "samwise", "gollum",
"legolas", "gandalf".}

\item{n}{Number of colors to return. If NULL, returns all colors in the palette.
If n is greater than the number of colors in the palette, colors will be
interpolated.}

\item{type}{Type of palette: "discrete" or "continuous". If NULL, will be
automatically determined based on n.}

\item{reverse}{Logical. Should the palette be reversed? Default is FALSE.}
}
\value{
A character vector of hex color codes
}
\description{
Returns a vector of colors from the specified LOTR palette.
}
\examples{
# Get all colors from the Shire palette
lotr_palette("shire")

# Get 5 colors from Mordor
lotr_palette("mordor", n = 5)

# Get reversed Rivendell palette
lotr_palette("rivendell", reverse = TRUE)

# Interpolate to get 20 colors
lotr_palette("minasmorgul", n = 20, type = "continuous")

}
