% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr_palette.R
\name{lotr_palette_names}
\alias{lotr_palette_names}
\title{Print all available LOTR palettes}
\usage{
lotr_palette_names(print_colors = TRUE)
}
\arguments{
\item{print_colors}{Logical. Should colors be displayed? Default is TRUE.}
}
\value{
Invisibly returns a character vector of palette names
}
\description{
Displays all available palette names and their colors
}
\examples{
# List all palettes
lotr_palette_names()

}
