% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mexpit}
\alias{mexpit}
\title{multi-expit (numerical overflow dealt with)}
\usage{
mexpit(v)
}
\arguments{
\item{v}{a vector of length K}
}
\value{
a vector of probabilities of length K; sums to one
}
\description{
multi-expit (numerical overflow dealt with)
}
\examples{
sigmoid <- function(v) mexpit(c(0,v))[2]
mexpit(c(1,0,0))
}
