% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{opt_colpermB}
\alias{opt_colpermB}
\title{Get optimal column permutation of a matrix B to match the matrix A}
\usage{
opt_colpermB(A, B)
}
\arguments{
\item{A}{a matrix (e.g., truth)}

\item{B}{another matrix (e.g., estimated); A and B must have the same dimensions;
B's columns may be permutated to best match those of A.}
}
\value{
a permutation of columns of B (represented by a permuted \code{1:ncol(A)})
}
\description{
Useful in simulations where the posterior sampling relabels the classes,
which is equivalent and cannot be told apart by the likelihood only.
}
\examples{

A <- matrix(c(1,2,3,4,5,6),nrow=2,ncol=3)
B <- A[,c(2,1,3)]
opt_colpermB(A,B) # should expect c(2,1,3)
}
