% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{permutations}
\alias{permutations}
\title{generate all permutations of 1 to n}
\usage{
permutations(n)
}
\arguments{
\item{n}{an integer}
}
\value{
a matrix of n columns and \code{factorial(n)} rows
}
\description{
Copied from Museful's solution from here
\url{https://stackoverflow.com/questions/11095992/generating-all-distinct-permutations-of-a-list-in-r}
This also mimics \code{RcppAlgos::permuteGeneral(n,n)}
}
\examples{
permutations(3)

}
