% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blca}
\alias{predict.blca}
\title{approximate predictive class probabilities for a new observation}
\usage{
\method{predict}{blca}(object, dat_pred = NULL, ...)
}
\arguments{
\item{object}{An \code{lcm_tree} class object; Output from \code{BayesLCA::blca()}}

\item{dat_pred}{a binary matrix with rows for observations and columns
for features (of the same dimension as in the training data)}

\item{...}{Other parameters}
}
\value{
a matrix with the same number of rows as \code{dat_pred} and
\code{K} columns (\code{K} is the number of latent classes);
each row correspond to a vector of predicted probabilities of
an observation in each class.
}
\description{
approximate predictive class probabilities for a new observation
}
