% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lcm_tree}
\alias{predict.lcm_tree}
\title{plots the groups discovered by lcm_tree on the tree for observations.}
\usage{
\method{predict}{lcm_tree}(object, dat_pred = NULL, leaf_ids = NULL, collapsed = TRUE, ...)
}
\arguments{
\item{object}{An \code{lcm_tree} class object; Output from \code{lcm_tree()}}

\item{dat_pred}{a binary matrix with rows for observations and columns
for features (of the same dimension as in the training data)}

\item{leaf_ids}{a vector of character strings, each being the name
of the leaf name for that observation}

\item{collapsed}{default to \code{TRUE}; if \code{FALSE}, no posterior median
model selection is done - so each leaf may have its own vector of
latent class proportions.}

\item{...}{Other parameters}
}
\value{
a matrix with the same number of rows as \code{dat_pred} and
\code{K} columns (\code{K} is the number of latent classes);
each row correspond to a vector of predicted probabilities of
an observation in each class.
}
\description{
plots the groups discovered by lcm_tree on the tree for observations.
}
