% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{print.lcm_tree}
\alias{print.lcm_tree}
\title{\code{print.lcm_tree} summarizes the results from \code{\link[=lcm_tree]{lcm_tree()}}.}
\usage{
\method{print}{lcm_tree}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link[=lcm_tree]{lcm_tree()}}.}

\item{...}{Arguments passed to summary and printing methods.}
}
\value{
Summary showing, for each group of leaf nodes discovered by \code{\link[=lotR]{lotR()}},
the classs prevalences, 95\\% credible intervals, number of leaf nodes in
per group, and number of observations per group.
}
\description{
\code{print.lcm_tree} summarizes the results from \code{\link[=lcm_tree]{lcm_tree()}}.
}
\seealso{
Other lcm_tree results: 
\code{\link{print.summary.lcm_tree_compact}()},
\code{\link{print.summary.lcm_tree_long}()},
\code{\link{summary.lcm_tree}()}
}
\concept{lcm_tree results}
