% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob2stick}
\alias{prob2stick}
\title{transform probability to stick proportions.}
\usage{
prob2stick(x)
}
\arguments{
\item{x}{a vector of probabilities (K)}
}
\value{
a vector K, with last element of 1; the elements are stick lengths in
the remaining part
}
\description{
transform probability to stick proportions.
}
