% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{rmse_bias_AB}
\alias{rmse_bias_AB}
\title{calculate root mean squared errors and biases}
\usage{
rmse_bias_AB(A, B)
}
\arguments{
\item{A}{a matrix (e.g., truth)}

\item{B}{another matrix (e.g., estimated); A and B must have the same dimensions;
B's columns must have already been permuted to best match those of A.}
}
\value{
a list
\describe{
\item{rmse_total}{root mean squared error for all entries}
\item{rmse_marg}{root mean squared error for all entries by column of A}
\item{frac_bias}{percent bias averaged over all entries}
\item{frac_bias_marg}{ percent bias by column of A}
}
}
\description{
used for comparing truth and estimates
}
\examples{

A <- matrix(c(1,2,3,4,5,6),nrow=2,ncol=3)
B <- A
rmse_bias_AB(A,B)
}
