% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_along_dim}
\alias{split_along_dim}
\title{create a list with members being the matrices along a specified dimension of an array}
\usage{
split_along_dim(a, n)
}
\arguments{
\item{a}{an array}

\item{n}{along which dimension to create a list}
}
\value{
a list
}
\description{
This is from a StackOverflow answer
}
\examples{

myarray <- array(c(1,2,3,4,5,6,7,8),c(2,2,2))

split_along_dim(myarray,1)
split_along_dim(myarray,2)
split_along_dim(myarray,3)
}
\seealso{
Other utility functions: 
\code{\link{make_list}()}
}
\concept{utility functions}
