% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{summary.lcm_tree}
\alias{summary.lcm_tree}
\title{\code{summary.lcm_tree} summarizes the results from \code{\link[=lcm_tree]{lcm_tree()}}.}
\usage{
\method{summary}{lcm_tree}(object, compact = FALSE, ...)
}
\arguments{
\item{object}{Output from \code{\link[=lcm_tree]{lcm_tree()}}.
An object of class "lcm_tree".# coeff_type Either "lcm_tree" or "ad_hoc"}

\item{compact}{If \code{TRUE}, a more compact summary of results is printed.
Only works well when the dimension of the variables is low
(say, < 4); otherwise the table takes up too much horizontal space.
Default is \code{FALSE}.}

\item{...}{Not used.}
}
\value{
see \code{\link[=print.lcm_tree]{print.lcm_tree()}}
}
\description{
Will have some randomness associated with the uppper and lower bounds,
because we simulated the Gaussian variables before converting to probabilities.
}
\seealso{
Other lcm_tree results: 
\code{\link{print.lcm_tree}()},
\code{\link{print.summary.lcm_tree_compact}()},
\code{\link{print.summary.lcm_tree_long}()}
}
\concept{lcm_tree results}
