% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_alpha_subid}
\alias{update_alpha_subid}
\title{Update the variational mean and variance for logit of
class-specific response probabilities (for the \code{s_u=1} component)}
\usage{
update_alpha_subid(
  u,
  g_phi,
  rmat,
  tau_1_t_u,
  h_pau,
  levels,
  E_eta,
  E_xi_u,
  subject_ids,
  v_lookup
)
}
\arguments{
\item{u}{node id}

\item{rmat}{a matrix of variational probabilities of all observations
belong to K classes; N by K; each row sums to 1}

\item{h_pau}{a numeric vector of length p indicating the branch length
between a node and its parent}

\item{levels}{a vector of possibly repeating integers from 1 to Fg, or L,}

\item{E_eta, }{E_xi the moments obtained from \code{\link{get_moments_cpp}};
E_eta are sums of ancestral elements of E_xi}

\item{subject_ids}{the ids of subjects in the leaf descendants of node u}

\item{v_lookup}{a vector of length equal to the total number of rows in X;
each element is an integer, indicating which leaf does the observation belong to.}

\item{tau_1_t}{variational Gaussian variances for alpha; matrix, dimension:(p by K-1)}

\item{X}{transformed data: 2Y-1}

\item{D_k_array, }{GvK_array, I_tilde,submat,idnotk utility quantities}
}
\value{
a list
\describe{
Named("resMu")=resMu, K-1 column vector

Named("resSigma")=resSigma, K-1 by K-1 matrix

}
}
\description{
Update the variational mean and variance for logit of
class-specific response probabilities (for the \code{s_u=1} component)
}
