% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_gamma_subid}
\alias{update_gamma_subid}
\title{Update the variational mean and variance for logit of
class-specific response probabilities (for the \code{s_u=1} component)}
\usage{
update_gamma_subid(
  u,
  g_psi,
  rmat,
  tau_2_t_u,
  h_pau,
  levels,
  E_beta,
  E_zeta_u,
  X,
  subject_ids,
  v_lookup
)
}
\arguments{
\item{u}{node id}

\item{g_psi}{g transformed psi}

\item{rmat}{a matrix of variational probabilities of all observations
belong to K classes; N by K; each row sums to 1}

\item{h_pau}{a numeric vector of length p indicating the branch length
between a node and its parent}

\item{levels}{a vector of possibly repeating integers from 1 to Fg, or L,
where L indicates the number of node subsets, each of which share the hyperparameters
\code{rho} - the prior probability of \code{s_u=1}, \code{tau_1[l,]} - K-1, \code{tau_2[l,,]}, J by K,
the prior variances for alpha and gamma.}

\item{E_beta, }{E_zeta the moments obtained from \code{\link{get_moments_cpp}};
E_beta are sums of ancestral elements of E_zeta}

\item{X}{transformed data: 2Y-1}

\item{subject_ids}{the ids of subjects in the leaf descendants of node u}

\item{v_lookup}{a vector of length equal to the total number of rows in X;
each element is an integer, indicating which leaf does the observation belong to.}

\item{psi}{a (pL,J,K) array of local variational parameters for approximating
the \code{expit(X^v_ij * beta^v_jk)}}

\item{tau_2_t}{variational Gaussian variances for gamma; array, dimension:(J,K,p)}
}
\value{
a list
\describe{
Named("resMu")=resMu, J by K

Named("resSigma")=resSigma, J by K

}
}
\description{
Update the variational mean and variance for logit of
class-specific response probabilities (for the \code{s_u=1} component)
}
