% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_hyperparams}
\alias{update_hyperparams}
\title{update hyperparameters}
\usage{
update_hyperparams(
  Y,
  A,
  outcomes_units,
  ancestors,
  h_pau,
  levels,
  v_units,
  X,
  n,
  J,
  p,
  pL,
  Fg,
  prob,
  mu_gamma,
  mu_alpha,
  rmat,
  sigma_gamma,
  Sigma_alpha,
  tau_1_t,
  tau_2_t,
  a_t,
  b_t,
  E_beta_sq,
  E_eta_sq,
  E_beta,
  E_eta,
  psi,
  g_psi,
  phi,
  g_phi,
  tau_1,
  tau_2,
  a,
  b,
  K,
  tau_update_levels,
  update_hyper
)
}
\arguments{
\item{Y}{outputs from
\code{design_tree()}}

\item{A}{outputs from
\code{design_tree()}}

\item{outcomes_units}{outputs from
\code{design_tree()}}

\item{ancestors}{outputs from
\code{design_tree()}}

\item{h_pau}{outputs from
\code{design_tree()}}

\item{levels}{outputs from
\code{design_tree()}}

\item{X}{computed from the data by \code{lcm_tree()}}

\item{n}{computed from the data by \code{lcm_tree()}}

\item{J}{computed from the data by \code{lcm_tree()}}

\item{p}{computed from the data by \code{lcm_tree()}}

\item{pL}{computed from the data by \code{lcm_tree()}}

\item{Fg}{computed from the data by \code{lcm_tree()}}

\item{prob}{variational parameters updated by \code{update_vi_params()}}

\item{mu_gamma}{variational parameters updated by \code{update_vi_params()}}

\item{mu_alpha}{variational parameters updated by \code{update_vi_params()}}

\item{rmat}{variational parameters updated by \code{update_vi_params()}}

\item{sigma_gamma}{variational parameters updated by \code{update_vi_params()}}

\item{Sigma_alpha}{variational parameters updated by \code{update_vi_params()}}

\item{tau_1_t}{variational parameters updated by \code{update_vi_params()}}

\item{tau_2_t}{variational parameters updated by \code{update_vi_params()}}

\item{a_t}{variational parameters updated by \code{update_vi_params()}}

\item{b_t}{variational parameters updated by \code{update_vi_params()}}

\item{psi}{parameters updated by \code{update_hyperparams()}}

\item{g_psi}{parameters updated by \code{update_hyperparams()}}

\item{phi}{parameters updated by \code{update_hyperparams()}}

\item{g_phi}{parameters updated by \code{update_hyperparams()}}

\item{tau_1}{parameters updated by \code{update_hyperparams()}}

\item{tau_2}{parameters updated by \code{update_hyperparams()}}

\item{a}{fixed hyperparameters}

\item{b}{fixed hyperparameters}

\item{K}{computed from the data by \code{lcm_tree()}}

\item{update_hyper}{Logical, \code{TRUE} or \code{FALSE} to indicate}
}
\value{
a list of updated hyperparameters: tau_1,tau_2,psi,g_psi,phi,g_phi,
along with a new ELBO value.
}
\description{
update hyperparameters
}
