% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_rmat}
\alias{update_rmat}
\title{Update the variational probabilities of each observation in one of K classes}
\usage{
update_rmat(
  psi,
  g_psi,
  phi,
  g_phi,
  X,
  E_beta,
  E_eta,
  E_beta_sq,
  E_eta_sq,
  v_lookup
)
}
\arguments{
\item{psi, g_psi, phi, g_phi}{local variational parameters}

\item{X}{transformed data: 2Y-1}

\item{E_beta, E_eta, E_beta_sq, E_eta_sq}{moment updates produced by \code{\link[=get_moments_cpp]{get_moments_cpp()}}}

\item{v_lookup}{a vector of length equal to the total number of rows in \code{X};
each element is an integer, indicating which leaf does the observation belong to.}
}
\value{
N by K variational multinomial probabilities; row sums are 1s.
}
\description{
This function updates the N by K matrix \code{rmat} in the package
}
