% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_rZ}
\alias{update_rZ}
\title{Z rmat update}
\usage{
update_rZ(
  psi,
  g_psi,
  phi,
  g_phi,
  X,
  E_beta,
  E_eta,
  E_beta_sq,
  E_eta_sq,
  v_lookup
)
}
\arguments{
\item{psi, }{g_psi, phi, g_phi local variational parameters}

\item{X}{transformed data: 2Y-1}

\item{E_beta, }{E_eta_diff, E_beta_sq, E_eta_diff_sq
intermediate moment updates produced by \code{\link{get_moments_cpp}}}

\item{v_lookup}{a vector of length equal to the total number of rows in X;
each element is an integer, indicating which leaf does the observation belong to.}

\item{idnotk}{utility matrix; column k indicates the indices that are not k}
}
\value{
n by K variational multinomial probabilities; row sums are 1s.
}
\description{
Z rmat update
}
