% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_vi_params}
\alias{update_vi_params}
\title{update variational parameters (core internal function)}
\usage{
update_vi_params(
  Y,
  A,
  outcomes_units,
  outcomes_nodes,
  ancestors,
  cardanc,
  v_units,
  h_pau,
  levels,
  X,
  n,
  J,
  p,
  pL,
  Fg,
  subject_id_list,
  prob,
  mu_gamma,
  mu_alpha,
  rmat,
  sigma_gamma,
  Sigma_alpha,
  tau_1_t,
  tau_2_t,
  a_t,
  b_t,
  psi,
  g_psi,
  phi,
  g_phi,
  tau_1,
  tau_2,
  a,
  b,
  K,
  s_u_zeroset,
  s_u_oneset
)
}
\arguments{
\item{Y, A, outcomes_units, outcomes_nodes, h_pau, ancestors, levels}{outputs from
\code{design_tree()}}

\item{X, n, J, p, pL, Fg, K, }{computed from the data by \code{lcm_tree()}}

\item{prob, mu_gamma, mu_alpha, rmat, sigma_gamma, Sigma_alpha, tau_1_t, tau_2_t, a_t, b_t, }{variational parameters updated by \code{update_vi_params()}}

\item{psi, g_psi, phi, g_phi, tau_1, tau_2, }{parameters updated by \code{update_hyperparams()}}

\item{a, b, K}{fixed hyperparameters}
}
\description{
NB: needs speed up for multiplications over multiple arrays/lists of matrices.
}
\concept{Internal VI functions}
