% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_tree.R
\name{design_tree}
\alias{design_tree}
\title{Organize the data around the edge-weighted rooted tree}
\usage{
design_tree(
  Y,
  leaf_ids,
  mytree,
  rootnode = "Node1",
  weighted_edge = FALSE,
  Z_obs = NULL
)
}
\arguments{
\item{Y}{N by J matrix of binary measurements}

\item{leaf_ids}{This is a character string}

\item{mytree}{the tree (an \code{igraph} object) that contains the node,
edge, edge-length information.}

\item{rootnode}{a character string that indicates the name of the root node.}

\item{weighted_edge}{logical: TRUE for incorporating the branch lengths -
then the mytree must have this info; if FALSE, every branch, including
an imaginary rootnode to past branch, is set to have length 1.}

\item{Z_obs}{Default is \code{NULL}, a two-column matrix of (id,class indicator)
for a subset of people; # rows = # rows of Y; 2nd column is \code{NA} if
a person's class indicator is unknown.
The rows will be reordered according to the reordered \code{Y}.}
}
\value{
A list of data and tree information
\itemize{
\item \code{Y} A matrix of n by J; binary measurements with rows ordered by
      leaf ("leaf_ids") groups (e.g., information from phylogenetic tree).
\item \code{A} A matrix of p by p; ancestor matrix
\item \code{A_leaves} A matrix of pL by p; ancestor matrix only for leaves
\item \code{leaf_ids} This is a vector of length n; ordered by the leaves as
             specified by the tree.
\item \code{leaf_ids_units} A list of length \code{p[L]}, each element
           is a numeric vector of subject ids belonging to each leaf ("leaf_ids")
\item \code{leaf_ids_nodes} a list of length \code{p}, each element
           is a numeric vector indicating the leaf nodes.
\item \code{ancestors} a list of length \code{p[L]},
     each element is the numeric vector of ancestors
     (including the leaf node as well).
\item \code{edge_lengths} a list of length \code{p[L]},
      each element is a numeric vector of edge lengths from the root node
      to the leaf.
\item \code{h_pau} a numeric vector of length \code{p}; each value
\item \code{v_units} a vector of length equal to the total number of rows in X;
each element is an integer, indicating which leaf does the observation belong to.
\item \code{subject_id_list} a list of length p; each element is a vector of subject ids
that are in the leaf descendants of node u (internal or leaf node)
\item \code{ord} the permutation to order the original rows to produce the final ordering
of the rows
is the length between the node and its parent; the root node has no parent
and the edge length is set to 1.
}
}
\description{
NB: currently this minimally built; need some checking functions
}
