% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_lcm_tree}
\alias{fit_lcm_tree}
\title{Fit latent class model with tree-structured shrinkage over the observations
(this function largely follows Thomas, E. \code{moretrees})}
\usage{
fit_lcm_tree(
  dsgn,
  vi_params_init,
  hyperparams_init,
  random_init,
  random_init_vals,
  tol,
  tol_hyper,
  max_iter,
  print_freq,
  quiet,
  plot_fig,
  shared_tau,
  update_hyper_freq,
  hyper_fixed
)
}
\arguments{
\item{dsgn}{a list of data and other information organized according to the tree}

\item{vi_params_init, hyperparams_init, random_init, random_init_vals, tol, tol_hyper, max_iter, print_freq, quiet, plot_fig, shared_tau, update_hyper_freq, hyper_fixed}{initial values and updating protocols. Explained more in the wrapper function \code{\link[=lcm_tree]{lcm_tree()}}}
}
\value{
a list with model updates; because the variational posterior
is comprised of familiar distributional forms that can be  determined
by the moments, the returned values are moments:

\describe{
\item{\code{vi_params}}{named list of final variational parameter estimates}
\item{\code{hyperparams}}{named list of final hyperparameter estimates}
\item{\code{hyper_fixed}}{named list of fixed hyperparameters}
\item{\code{ELBO_track}}{numeric vector containing the values of the objective function
(ELBO) at the end of every iteration}
}
}
\description{
Fit latent class model with tree-structured shrinkage over the observations
(this function largely follows Thomas, E. \code{moretrees})
}
\concept{internal VI functions}
