% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_moments_cpp}
\alias{get_moments_cpp}
\title{Calculate variational moments during the updates}
\usage{
get_moments_cpp(
  prob,
  mu_gamma,
  sigma_gamma,
  mu_alpha,
  Sigma_alpha,
  anc,
  cardanc
)
}
\arguments{
\item{prob}{variational probabilities for \code{s_u}; length p}

\item{mu_gamma}{variational Gaussian means (for \code{s_u=1} component) for J*K
logit(class-specific response probabilities); (J,K,p) array; In R, we used a list of p (J,K) matrices}

\item{sigma_gamma}{variational Gaussian variances (for \code{s_u=1} component)
for J*K logit(class-specific response probabilities); (J,K,p) array}

\item{mu_alpha}{variational Gaussian mean vectors (for \code{s_u=1} component) -
this is a p by K-1 matrix; in R, we used a list of p vectors (each of length K-1)}

\item{Sigma_alpha}{variational Gaussian variances (for \code{s_u=1} component)
- this is an array of dimension (K-1, K-1, p); in R, we used a list of p matrices,
each of dimension K-1 by K-1.}

\item{anc}{a list of pL vectors, each vector has the node ids of the ancestors;
lengths may differ. The ancestors include the node concerned.}

\item{cardanc}{a numeric vector of length pL; integers. The number
of ancestors for each leaf node}
}
\value{
a List

\describe{
  return List::create(Named("E_beta")=E_beta,
   Named("E_beta_sq")=E_beta_sq,
   Named("E_eta")=E_eta,
   Named("E_eta_sq")=E_eta_sq);
}
}
\description{
Get all moments that need update when iterating over a total of p internal and leaf nodes
}
\examples{
prob = 0.2
mu1 = 2
mu0 = 0

sigma1 = 3
sigma0 = 1

N = 1000000
s = rbinom(N,1,prob)
y <- rep(NA,length(s))
for (i in seq_along(s)){
      y[i] <- rnorm(1,mu1,sqrt(sigma1))*s[i] +
        rnorm(1,mu0,sqrt(sigma0))*(1-s[i])
}

var(s*y)
prob*(sigma1+(1-prob)*mu1^2)


mean(y^2)
prob*(sigma1+mu1^2) + (1-prob)*sigma0

mean(s*y^2)
prob*(sigma1+mu1^2)

mean(s*y)
prob*mu1
}
