% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{lcm_by_group}
\alias{lcm_by_group}
\title{estimate latent class models by pre-specified groups of observations,
e.g., the ones obtained from the tree-shrunk LCM.}
\usage{
lcm_by_group(Y, outcomes, members_list, K, ...)
}
\arguments{
\item{Y}{data}

\item{outcomes}{individual's leaf indicators in a vector}

\item{members_list}{a list, of length G, which is the number of unique
groups of leaves}

\item{K}{the number of classes}

\item{...}{other arguments for \code{\link[BayesLCA]{blca}}}
}
\description{
estimate latent class models by pre-specified groups of observations,
e.g., the ones obtained from the tree-shrunk LCM.
}
