% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotR.R
\name{lotr}
\alias{lotr}
\title{Lord of the Rings soundboard}
\usage{
lotr(clip = "over")
}
\arguments{
\item{clip}{Choose soundclip.  Defaults to "over". Possible values: "clever", "mellon", "no", "over", "shallnotpass", "shut", "task", "walls", "wizard", "work"}
}
\description{
Plays a quote from LotR upon execution
}
\examples{
lotr("mellon")
}
\keyword{Lotr}
