% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{lotR_blca}
\alias{lotR_blca}
\title{Simulate data and subject-specific indicators from latent class models}
\usage{
lotR_blca(n, itemprob = 0.5, classprob = 1, fit = NULL)
}
\arguments{
\item{n}{sample size}

\item{itemprob}{item probabilities}

\item{classprob}{class probabilities}

\item{fit}{fitted object}
}
\value{
a list; \code{x}: data; \code{z} a vector of integer indicators
of class membership.
}
\description{
Simulate data and subject-specific indicators from latent class models
}
\seealso{
\code{\link[BayesLCA:blca]{BayesLCA::blca()}}
}
