% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lower_bd}
\alias{lower_bd}
\title{lower bound of expit(x)}
\usage{
lower_bd(xi)
}
\arguments{
\item{xi}{local variational parameter, positive value, this is where
the \code{expit(xi) = lower_bd(xi)}}
}
\value{
a positive value between 0 and 1
}
\description{
lower bound is quardratic in the exponent
}
\examples{

par(mfrow=c(2,3))
xi_seq <- c(0.1,0.5,1,2,4,6)
  x <- seq(-5,5,by=0.1)
for (i in 1:length(xi_seq)){
  xi <- xi_seq[i]
  y1 <- lower_bd(xi)(x)
  y2 <- lower_bd(-xi)(x)

  plot(x,y1,type="l",col="red",ylim=c(0,1),main=paste0("xi= ",xi))
  points(x,y2,type="l",col="blue")
  points(x,1/(1+exp(-x)),type="l",col="orange")
  abline(v=c(-xi,xi),col="gray")
  legend("topleft",c("expit","bound"),col=c("orange","blue"),lty=c(1,1))
}

}
\seealso{
Other VI functions: 
\code{\link{g_fun.vec}()},
\code{\link{g_fun0}()},
\code{\link{g_fun}()}
}
\concept{VI functions}
