% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.lcm_tree}
\alias{plot.lcm_tree}
\title{plots the groups discovered by lcm_tree on the tree for observations.}
\usage{
\method{plot}{lcm_tree}(
  x,
  group.text.size = 4,
  group.text.offset = 0.1,
  legend.text.size = 10,
  layout = "rectangular",
  horizontal = TRUE,
  ...
)
}
\arguments{
\item{x}{Output from \code{moretrees()}
An object of class "moretrees_result".}

\item{group.text.size}{Text size for the group labels}

\item{group.text.offset}{Offset of the group label from the
leaves of the tree}

\item{legend.text.size}{Text size for legend}

\item{layout}{Layout for the tree, most likely "rectangular" (the default)
or "slanted", but see the \code{layout} option of \code{ggtree()} for more
possibilities}

\item{horizontal}{If TRUE (the default), the tree will be plotted with
the root node at the top and all other nodes below. If FALSE, the tree
will be plotted with the root node to the left and all other nodes to
the right.}

\item{...}{Not used.}
}
\value{
A plot showing the groups discovered by MOReTreeS on the original
leaf_ids tree.
}
\description{
plots the groups discovered by lcm_tree on the tree for observations.
}
\examples{
# See vignette

}
