% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_est_resp}
\alias{plot_est_resp}
\title{plots the estimated response probabilities by class}
\usage{
plot_est_resp(x, xlab_nm = "feature")
}
\arguments{
\item{x}{An \code{lcm_tree} class object; Output from \code{lcm_tree()}}

\item{xlab_nm}{x axis name; a character string}
}
\value{
A plot showing the groups discovered by MOReTreeS on the original
leaf_ids tree.
}
\description{
This estimated by \code{\link[=lcm_tree]{lcm_tree()}}
}
\examples{
# See vignette

}
