% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{rmse_bias_AB}
\alias{rmse_bias_AB}
\title{Get optimal column permutation of a matrix B to match the matrix A}
\usage{
rmse_bias_AB(A, B)
}
\arguments{
\item{A}{a matrix (e.g., truth)}

\item{B}{another matrix (e.g., estimated); A and B must have the same dimensions;
B's columns may be permutated to best match those of A.}
}
\value{
a list
\describe{
\item{rmse_total}{root mean squared error for all entries}
\item{rmse_marg}{root mean squared error for all entries by column of A}
\item{frac_bias}{root mean squared error for all entries}
\item{frac_bias_marg}{root mean squared error for all entries by column of A}
}
}
\description{
Useful in simulations where the posterior sampling relabels the classes,
which is equivalent and cannot be told apart by the likelihood only.
}
\examples{

A <- matrix(c(1,2,3,4,5,6),nrow=2,ncol=3)
B <- A
rmse_bias_AB(A,B)
}
