% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_hyperparams}
\alias{update_hyperparams}
\title{update hyperparameters in the latent class model with observations organized in
a tree}
\usage{
update_hyperparams(
  Y,
  A,
  leaf_ids_units,
  ancestors,
  h_pau,
  levels,
  v_units,
  X,
  n,
  J,
  p,
  pL,
  Fg,
  prob,
  prob_gamma,
  mu_gamma,
  mu_alpha,
  rmat,
  sigma_gamma,
  Sigma_alpha,
  tau_1_t,
  tau_2_t,
  a_t,
  b_t,
  E_beta_sq,
  E_eta_sq,
  E_beta,
  E_eta,
  psi,
  g_psi,
  phi,
  g_phi,
  tau_1,
  tau_2,
  shared_tau,
  a,
  b,
  K,
  tau_update_levels,
  update_hyper,
  quiet
)
}
\arguments{
\item{Y}{outputs from
\code{\link{design_tree}}, which reorders the nodes by internal and leaf nodes; the observations are also
ordered from low- to high-indexed leaf nodes.}

\item{A}{outputs from
\code{\link{design_tree}}, which reorders the nodes by internal and leaf nodes; the observations are also
ordered from low- to high-indexed leaf nodes.}

\item{leaf_ids_units}{outputs from
\code{\link{design_tree}}, which reorders the nodes by internal and leaf nodes; the observations are also
ordered from low- to high-indexed leaf nodes.}

\item{ancestors}{outputs from
\code{\link{design_tree}}, which reorders the nodes by internal and leaf nodes; the observations are also
ordered from low- to high-indexed leaf nodes.}

\item{h_pau}{outputs from
\code{\link{design_tree}}, which reorders the nodes by internal and leaf nodes; the observations are also
ordered from low- to high-indexed leaf nodes.}

\item{levels}{outputs from
\code{\link{design_tree}}, which reorders the nodes by internal and leaf nodes; the observations are also
ordered from low- to high-indexed leaf nodes.}

\item{v_units}{outputs from
\code{\link{design_tree}}, which reorders the nodes by internal and leaf nodes; the observations are also
ordered from low- to high-indexed leaf nodes.}

\item{X}{computed from the data by \code{\link{lcm_tree}} at the beginning before the VI updates}

\item{n}{computed from the data by \code{\link{lcm_tree}} at the beginning before the VI updates}

\item{J}{computed from the data by \code{\link{lcm_tree}} at the beginning before the VI updates}

\item{p}{computed from the data by \code{\link{lcm_tree}} at the beginning before the VI updates}

\item{pL}{computed from the data by \code{\link{lcm_tree}} at the beginning before the VI updates}

\item{Fg}{computed from the data by \code{\link{lcm_tree}} at the beginning before the VI updates}

\item{prob}{variational parameters updated by \code{\link{update_vi_params}}}

\item{prob_gamma}{variational parameters updated by \code{\link{update_vi_params}}}

\item{mu_gamma}{variational parameters updated by \code{\link{update_vi_params}}}

\item{mu_alpha}{variational parameters updated by \code{\link{update_vi_params}}}

\item{rmat}{variational parameters updated by \code{\link{update_vi_params}}}

\item{sigma_gamma}{variational parameters updated by \code{\link{update_vi_params}}}

\item{Sigma_alpha}{variational parameters updated by \code{\link{update_vi_params}}}

\item{tau_1_t}{variational parameters updated by \code{\link{update_vi_params}}}

\item{tau_2_t}{variational parameters updated by \code{\link{update_vi_params}}}

\item{a_t}{variational parameters updated by \code{\link{update_vi_params}}}

\item{b_t}{variational parameters updated by \code{\link{update_vi_params}}}

\item{E_beta_sq, E_eta_sq, E_beta, E_eta}{moments computed by \code{\link{update_vi_params}}}

\item{psi}{parameters updated by \code{update_hyperparams()}}

\item{g_psi}{parameters updated by \code{update_hyperparams()}}

\item{phi}{parameters updated by \code{update_hyperparams()}}

\item{g_phi}{parameters updated by \code{update_hyperparams()}}

\item{tau_1}{parameters updated by \code{update_hyperparams()}}

\item{tau_2}{parameters updated by \code{update_hyperparams()}}

\item{shared_tau}{parameters updated by \code{update_hyperparams()}}

\item{a}{fixed hyperparameters not to be updated.}

\item{b}{fixed hyperparameters not to be updated.}

\item{K}{fixed hyperparameters not to be updated.}

\item{tau_update_levels}{a numeric vector, specifies which levels of hyperparameters to update}

\item{update_hyper}{Logical, \code{TRUE} or \code{FALSE} to indicate
whether to update \code{tau_1} and \code{tau_2}. This is computed at every iteration
in \code{\link{fit_lcm_tree}}}

\item{quiet}{default to \code{FALSE}, which prints intermediate updates of hyperparameters}
}
\value{
a list of updated hyperparameters: tau_1,tau_2,psi,g_psi,phi,g_phi,
along with a new ELBO value.
}
\description{
update hyperparameters in the latent class model with observations organized in
a tree
}
